Add support to some ranges. #114
This adds quick camelcasing names for issue  #114
* U+2900-U+297F	Supplemental Arrows-B
* U+2980-U+29FF	Miscellaneous Mathematical Symbols-B
* U+2A00-U+2AFF	Supplemental Mathematical Operators
* U+2B00-U+2BFF	Miscellaneous Symbols and Arrows

This also removes 2 gremlin characters from the "Supplementary" word in the unicodePlaneNames. I can't see if they caused any harm, but they should not be there.

This also adds "Combining Diacritical Marks for Symbols", as the module was there but not committed. Basic camelcase names seem best. But noted that the "Combining Diacritical Marks Supplement" does everything lowercase and consistency is hard.