Fix issue  #113
Arabc AIN has a conflict with "LATIN LETTER AIN" from Phonetic Extensions. This causes the Arabic AIN to get a prefix and then it loses its connection with the positional forms.
This forces a script prefix on the AIN from Phonetic Extensions, as less depends on that name. As far as I can tell.